(m_cheat.c) Silence warnings

+++ src/r_data.c
@@ -499,7 +499,7 @@ void R_PrecacheLevel(void)
 
   {
     size_t size = numflats > numsprites  ? numflats : numsprites;
-    hitlist = Z_Malloc(((size_t)numtextures > size) ? numtextures : size, PU_STATIC, 0);
+    hitlist = Z_Malloc(((size_t)numtextures > size) ? (size_t)numtextures : size, PU_STATIC, 0);
   }
 
   // Precache flats.
